+++ crates/swc_ecma_transforms_module/tests/paths/issue-4585/output/index.ts
@@ -1,0 +1,1 @@
+import "./src/feat.js";